impr: Replace `std::time::Instant` with `instant::Instant`
Replace `std::time::Instant` with `instant::Instant`.
`instant::Instant`, when compiled natively, is just a type alias for
`std::time::Instant`, which means that this change has no effect on
native platforms.

On platforms like `wasm32-unknown-unknown`, the `instant` crate can be
compiled with either the `instant/stdweb` or `instant/wasm-bindgen`
features. When compiled with these features, `instant::Instant` emulates
`std::time::Instant` using features like `performance.now()`. This
allows `instant` (and transitively `tessellation`) to run on WASM
platforms.

To compile `tessellation` with support for WASM platforms, enable either
the `stdweb` or `wasm-bindgen` feature of `tessellation`. These features
enable the respective `stdweb` or `wasm-bindgen` features of `instant`.
Whether you should use `stdweb` or `wasm-bindgen` depends on which
of these the rest of your application/library uses.

Even if the `stdweb` or `wasm-bindgen` feature is enabled, the `instant`
crate will continue to rely on `std::time::Instant` as long as your are
not targetting wasm32. This allows for portable code that works on both
native and WASM platforms. To support this behaviour, the `wasm-bindgen`
feature is enabled by default. The choice of `wasm-bindgen` over
`stdweb` was arbitrary.